feature: Rename useSearchFilters to useManageQueryParams hook
